backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yt-dlp>=2023.12.30
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
